+++ src/utils/validators.js
@@ -1,14 +1,14 @@
-const repoRegex = /https?:\/\/github.com\/(\w+)\/(\w+)/;
+const repoRegex = /https?:\/\/github.com\/([\w-_]+)\/([\w-_]+)/;
 
 const urlMatch = (url) => (
-  url && url.match(repoRegex)
+  url && url.match(repoRegex) || null
 );
 
 /**
  * Splits URL to owner and repo name.
  */
 const splitUrl = (url) => (
-  urlMatch(url).slice(1, 3)
+  urlMatch(url) && urlMatch(url).slice(1, 3)
 );
 
 export {