+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Change Log
 
+## [2020.12.14]
+  - Prevent submit on empty input
+  - Improve repo input validation regex
+  - Handle search error, refs #6
+
 ## [2020.12.10]
   - Initial UI release